Major renaming action, some tweaks and ifs

+++ lab1/ex06.sh
@@ -5,7 +5,7 @@ directory2_name=$2
 
 for file in $directory1_name/*
 do
-	if [ -x $file  ]
+	if [ -x $file -a -f $file  ]
 	then
 		mv $file $directory2_name
 	fi
